WEB-33 ACTION-01 ACTION-04 CONTENT-04 CONTENT-05 CONTENT-06 CONTENT-07 ARCH-02 ARCH-03 EEAT-02 PERF-01: update homepage content and loading
