+++ src/app/globals.css
@@ -1440,6 +1440,12 @@ button { font: inherit; color: inherit; background: none; border: 0; cursor: poi
   font-size: 15px;
   line-height: 1.45;
 }
+.nlg-attestation-status .nlg-attestation-methodology {
+  max-width: 70ch;
+  margin-top: 10px;
+  color: rgba(199, 210, 219, 0.58);
+  font-size: 13px;
+}
 .nlg-attestation-meta-strip {
   margin-top: 12px;
   font-family: var(--font-jetbrains-mono), ui-monospace, monospace;
@@ -1447,7 +1453,7 @@ button { font: inherit; color: inherit; background: none; border: 0; cursor: poi
   line-height: 1.4;
   letter-spacing: 0.12em;
   color: rgba(199, 210, 219, 0.42);
-  text-transform: uppercase;
+  text-transform: none;
 }
 .nlg-attestation-dots {
   width: min(100%, 720px);
@@ -2066,6 +2072,22 @@ button { font: inherit; color: inherit; background: none; border: 0; cursor: poi
   border-radius: 12px;
   background: var(--bg-2);
 }
+.nlg-form-placeholder {
+  min-height: 760px;
+  justify-content: stretch;
+}
+.nlg-form-placeholder span {
+  display: block;
+  min-height: 48px;
+  border: 1px solid var(--line-2);
+  border-radius: 6px;
+  background:
+    linear-gradient(90deg, rgba(255, 255, 255, 0.025), rgba(0, 174, 239, 0.05), rgba(255, 255, 255, 0.025)),
+    var(--bg-1);
+}
+.nlg-form-placeholder span:nth-child(5) {
+  min-height: 120px;
+}
 .nlg-form-row {
   display: grid;
   grid-template-columns: 1fr 1fr;
@@ -2227,7 +2249,18 @@ button { font: inherit; color: inherit; background: none; border: 0; cursor: poi
   margin: 0 0 16px;
   text-wrap: balance;
 }
-.nlg-cta-text .nlg-dim { color: var(--fg-3); font-weight: 500; font-size: 0.78em; letter-spacing: -0.015em; display: block; margin-top: 16px; max-width: 30ch; }
+.nlg-cta-subtext {
+  position: relative;
+  font-family: var(--font-montserrat), system-ui, sans-serif;
+  font-size: clamp(23px, 2.95vw, 40px);
+  line-height: 1.15;
+  letter-spacing: -0.015em;
+  font-weight: 500;
+  color: var(--fg-3);
+  max-width: 30ch;
+  margin: 0 0 16px;
+  text-wrap: balance;
+}
 .nlg-cta-actions { position: relative; display: flex; align-items: center; gap: 20px; margin-top: 40px; flex-wrap: wrap; }
 
 /* ───────── Footer ───────── */
@@ -2238,7 +2271,7 @@ button { font: inherit; color: inherit; background: none; border: 0; cursor: poi
 }
 .nlg-footer-grid {
   display: grid;
-  grid-template-columns: 1.2fr 1fr 1fr 1fr;
+  grid-template-columns: 1.35fr 1fr 1fr;
   gap: 48px;
   margin-bottom: 48px;
 }
@@ -2246,7 +2279,9 @@ button { font: inherit; color: inherit; background: none; border: 0; cursor: poi
   .nlg-footer-grid { grid-template-columns: 1fr 1fr; gap: 32px; }
 }
 .nlg-footer-brand { display: flex; align-items: center; gap: 10px; margin-bottom: 16px; font-weight: 600; }
-.nlg-footer-tag { font-size: 13px; color: var(--fg-3); line-height: 1.5; max-width: 32ch; margin: 0 0 20px; }
+.nlg-footer-tag,
+.nlg-footer-reach { font-size: 13px; color: var(--fg-3); line-height: 1.5; max-width: 32ch; margin: 0; }
+.nlg-footer-reach { margin-top: 8px; margin-bottom: 20px; }
 .nlg-footer-social {
   display: inline-flex;
   align-items: center;
@@ -2348,12 +2383,57 @@ button { font: inherit; color: inherit; background: none; border: 0; cursor: poi
 .nlg-nav.nlg-scrolled .nlg-nav-in { height: 56px; }
 .nlg-nav-brand { display: flex; align-items: center; gap: 10px; font-weight: 600; font-size: 14px; letter-spacing: -0.01em; color: var(--fg-1); }
 .nlg-nav-links { display: flex; gap: 28px; list-style: none; margin: 0; padding: 0; }
+.nlg-nav-item {
+  position: relative;
+}
 .nlg-nav-links a {
   font-size: 14px;
   color: var(--fg-2);
   transition: color var(--dur-fb) var(--ease-fb);
 }
 .nlg-nav-links a:hover { color: var(--fg-1); }
+.nlg-nav-submenu {
+  position: absolute;
+  top: calc(100% + 14px);
+  left: -16px;
+  min-width: 232px;
+  padding: 10px;
+  display: grid;
+  gap: 4px;
+  border: 1px solid var(--line-2);
+  border-radius: 10px;
+  background: rgba(5, 16, 31, 0.96);
+  box-shadow: 0 24px 60px -36px rgba(0, 174, 239, 0.36);
+  opacity: 0;
+  pointer-events: none;
+  transform: translateY(-4px);
+  transition: opacity var(--dur-fb) var(--ease-fb), transform var(--dur-fb) var(--ease-fb);
+}
+.nlg-nav-submenu::before {
+  content: "";
+  position: absolute;
+  left: 0;
+  right: 0;
+  top: -16px;
+  height: 16px;
+}
+.nlg-nav-has-menu:hover .nlg-nav-submenu,
+.nlg-nav-has-menu:focus-within .nlg-nav-submenu {
+  opacity: 1;
+  pointer-events: auto;
+  transform: translateY(0);
+}
+.nlg-nav-submenu a {
+  padding: 9px 10px;
+  border-radius: 6px;
+  color: var(--fg-3);
+  font-size: 13px;
+}
+.nlg-nav-submenu a:hover,
+.nlg-nav-submenu a:focus-visible {
+  color: var(--fg-1);
+  background: rgba(0, 174, 239, 0.06);
+}
 .nlg-nav-right { display: flex; align-items: center; gap: 12px; }
 
 .nlg-nav-mobile-toggle {
@@ -2389,12 +2469,22 @@ button { font: inherit; color: inherit; background: none; border: 0; cursor: poi
   flex-direction: column;
   gap: 4px;
 }
+.nlg-mobile-nav .nlg-mobile-subnav {
+  padding: 0 0 8px 16px;
+  gap: 0;
+  border-left: 1px solid var(--line-2);
+}
 .nlg-mobile-nav li a {
   display: block;
   padding: 12px 4px;
   font-size: 16px;
   font-weight: 500;
   color: var(--fg-2);
+}
+.nlg-mobile-nav .nlg-mobile-subnav a {
+  padding: 8px 4px;
+  font-size: 13px;
+  color: var(--fg-3);
 }
 .nlg-mobile-nav li a:hover { color: var(--fg-1); }
 .nlg-mobile-cta { padding-top: 16px; }
